Version 2.1.5: ~ Changed manageServer attribute to either admin_only or staff+ ~ Moved CmdSettings.java from guild to tickets
Listener:
~ Changed -test to !create
~ Users with staff role and/or manage server perm can now create multiple tickets
~ Moved checks for sending msg, reading chat, etc. to prevent errors, when bot can embed links, but not send msg
~ !create gets removed from the message

Utils:
~ The bot now removes the "user pinned a message" message

+++ src/main/java/com/andre601/suggesto/listener/CommandListener.java
@@ -3,12 +3,14 @@ package com.andre601.suggesto.listener;
 import com.andre601.suggesto.Supporto;
 import com.andre601.suggesto.utils.Database;
 import com.andre601.suggesto.utils.EmbedUtil;
+import com.andre601.suggesto.utils.LinkUtil;
 import com.andre601.suggesto.utils.PermUtil;
 import com.github.rainestormee.jdacommand.Command;
 import com.github.rainestormee.jdacommand.CommandHandler;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.entities.Guild;
 import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.entities.Role;
 import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
@@ -25,6 +27,17 @@ public class CommandListener extends ListenerAdapter {
     );
 
     private final CommandHandler HANDLER;
+
+    private Role getStaffRole(Guild guild){
+        Role role;
+        try{
+            role = guild.getRoleById(Database.getRoleID(guild));
+        }catch (Exception ex){
+            role = null;
+        }
+
+        return role;
+    }
 
     private TextChannel getSupportChannel(Guild guild){
         TextChannel channel;
@@ -62,6 +75,11 @@ public class CommandListener extends ListenerAdapter {
 
                     String prefix = Database.getPrefix(guild);
                     String raw = msg.getContentRaw();
+
+                    if(!PermUtil.canSeeChannel(tc)) return;
+                    if(!PermUtil.canSeeHistory(tc)) return;
+                    if(!PermUtil.canSendMsg(tc)) return;
+
                     if(!PermUtil.canEmbedLinks(tc)) {
                         tc.sendMessage(MessageFormat.format(
                                 "{0} I need permission to embed links in this channel!",
@@ -84,6 +102,8 @@ public class CommandListener extends ListenerAdapter {
                     String[] split = raw.split("\\s+", 2);
                     String commandString;
 
+                    Role role = getStaffRole(guild);
+
                     try{
                         if(raw.startsWith(prefix))
                             commandString = split[0].substring(prefix.length());
@@ -95,26 +115,63 @@ public class CommandListener extends ListenerAdapter {
                     Command command = HANDLER.findCommand(commandString.toLowerCase());
                     if(command == null) return;
                     if(command.hasAttribute("owner") && !PermUtil.isOwner(msg)) return;
-                    if(!PermUtil.canSeeChannel(tc)) return;
-                    if(!PermUtil.canSeeHistory(tc)) return;
-                    if(!PermUtil.canSendMsg(tc)) return;
                     if(!PermUtil.canAddReaction(tc)){
                         EmbedUtil.sendError(msg, "I need permissions, to add reactions!");
                         return;
                     }
-                    if(command.hasAttribute("manageServer") && !PermUtil.isAdmin(msg)){
-                        EmbedUtil.sendError(msg,"You need the permission `manage server` for this command!");
-                        return;
+                    if(command.hasAttribute("admin_only")){
+                        if(!PermUtil.isAdmin(msg)) {
+                            EmbedUtil.sendError(
+                                    msg,
+                                    "You need the permission `manage server` for this command!"
+                            );
+                            return;
+                        }
                     }
-                    if(!PermUtil.canManageMsg(tc)){
-                        EmbedUtil.sendError(msg, "I need permission to manage Messages!");
-                        return;
+                    if(command.hasAttribute("staff+")){
+                        if(role != null) {
+                            if (!PermUtil.isStaff(msg.getMember(), role)) {
+                                EmbedUtil.sendError(
+                                        msg,
+                                        "You need to have Staff-role or `manage server` permission"
+                                );
+                                return;
+                            }else
+                            if(PermUtil.isAdmin(tc, msg.getMember())){
+                                EmbedUtil.sendError(
+                                        msg,
+                                        "You need to have Staff-role or `manage server` permission"
+                                );
+                                return;
+                            }
+                        }else
+                        if(PermUtil.isAdmin(tc, msg.getMember())){
+                            EmbedUtil.sendError(
+                                    msg,
+                                    "You need to have Staff-role or `manage server` permission"
+                            );
+                            return;
+                        }
                     }
 
                     try {
                         HANDLER.execute(command, msg, split.length > 1 ? split[1] : "");
-                        msg.delete().queue();
+                        if(PermUtil.canManageMsg(tc)) msg.delete().queue();
                     }catch (Exception ex){
+                        EmbedUtil.sendError(
+                                msg,
+                                String.format(
+                                        "There was an issue performing this command!\n" +
+                                        "Please join the [supporto-guild](%s) and let the owner of the bot know " +
+                                        "about this issue!\n" +
+                                        "\n" +
+                                        "Cause of error:\n" +
+                                        "`%s`",
+                                        LinkUtil.INVITE_GUILD,
+                                        ex.getMessage()
+
+                                )
+                        );
                         Supporto.getLogger().error("Error in command", ex);
                     }
                 }
